+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Add `bin` to `ApplePayCard`
 * Add `submitForSettlement` to `SubscriptionGateway.retryCharge`
 * Add `options` -> `paypal` -> `description` for creating and updating subscriptions
+* Add `Dispute#accept`
+* Add `Dispute#find`
 
 ## 2.72.1